+++ Sources/SystemNotification/View+Geometry.swift
@@ -9,16 +9,9 @@
 import SwiftUI
 
 extension View {
-    
+
     /**
-     Bind the view's safe area to a binding.
-     */
-    func bindSafeAreaInsets(to binding: Binding<EdgeInsets>) -> some View {
-        background(safeAreaBindingView(for: binding))
-    }
-    
-    /**
-    Bind the view's size to a binding.
+     Bind the view's size to a binding.
     */
     func bindSize(to binding: Binding<CGSize>) -> some View {
         background(sizeBindingView(for: binding))
@@ -29,17 +22,6 @@ private extension View {
     
     func changeStateAsync(_ action: @escaping () -> Void) {
         DispatchQueue.main.async(execute: action)
-    }
-    
-    func safeAreaBindingView(for binding: Binding<EdgeInsets>) -> some View {
-        GeometryReader { geo in
-            self.safeAreaBindingView(for: binding, geo: geo)
-        }
-    }
-    
-    func safeAreaBindingView(for binding: Binding<EdgeInsets>, geo: GeometryProxy) -> some View {
-        changeStateAsync { binding.wrappedValue = geo.safeAreaInsets }
-        return Color.clear
     }
     
     func sizeBindingView(for binding: Binding<CGSize>) -> some View {